Rename removeWhitespace test helper -> removeExcessWhitespace
