+++ test-unit/test-helpers/remove-excess-whitespace.js
